# openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Task API",
		"description": "A simple CRUD Task API built with Express.",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000"
		}
	],
	"paths": {
		"/": {
			"get": {
				"summary": "API information",
				"responses": {
					"200": {
						"description": "API information returned successfully"
					}
				}
			}
		},
		"/health": {
			"get": {
				"summary": "Health check",
				"responses": {
					"200": {
						"description": "Server is healthy"
					}
				}
			}
		},
		"/tasks": {
			"get": {
				"summary": "Get all tasks",
				"parameters": [
					{
						"name": "done",
						"in": "query",
						"required": false,
						"description": "Optional completion filter. Use true to return completed tasks or false to return open tasks.",
						"schema": {
							"type": "string",
							"enum": ["true", "false"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "A list of tasks",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Task"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Create a new task",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewTask"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Task created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input"
					}
				}
			}
		},
		"/task/{id}": {
			"get": {
				"summary": "Get a task by ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Task found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					},
					"404": {
						"description": "Task not found"
					}
				}
			},
			"put": {
				"summary": "Update a task",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTask"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Task updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request"
					},
					"404": {
						"description": "Task not found"
					}
				}
			},
			"delete": {
				"summary": "Delete a task",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Task deleted successfully"
					},
					"404": {
						"description": "Task not found"
					}
				}
			}
		},
		"/stats": {
			"get": {
				"summary": "Get task statistics",
				"responses": {
					"200": {
						"description": "Task statistics returned successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Stats"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Task": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"title": {
						"type": "string",
						"example": "Buy milk"
					},
					"done": {
						"type": "boolean",
						"example": false
					}
				}
			},
			"NewTask": {
				"type": "object",
				"required": ["title"],
				"properties": {
					"title": {
						"type": "string",
						"example": "Buy milk"
					}
				}
			},
			"UpdateTask": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"example": "Buy bread"
					},
					"done": {
						"type": "boolean",
						"example": true
					}
				}
			},
			"Stats": {
				"type": "object",
				"properties": {
					"total": {
						"type": "integer",
						"example": 3
					},
					"done": {
						"type": "integer",
						"example": 2
					},
					"open": {
						"type": "integer",
						"example": 1
					}
				}
			}
		}
	}
}